타이핑 스크립트 버그 수정
